Gestión de usuarios + Administración
Funcionalidad:
- Acceso a administrador y reclutador.
- Acceso a usuarios y usuario por id.
- Acceso a perfiles y perfil por id.
- Implementación de Decorator authorize_user.
- Administración de test psicológicos.
